+++ components/@dopt/core/theme/src/types.ts
@@ -1,0 +1,11 @@
+export type StyleTheme =
+  | (string & {
+      className: string;
+      selector: string;
+    })
+  | null
+  | undefined;
+
+export interface StyleProps {
+  theme?: StyleTheme;
+}