+++ components/@dopt/react/tour/src/tour/dismiss-icon.tsx
@@ -1,0 +1,72 @@
+import * as classes from '../styles';
+import { classNameRoot } from '../const';
+
+import {
+  type ForwardedRef,
+  forwardRef,
+  type MouseEventHandler,
+  type ComponentPropsWithRef,
+} from 'react';
+
+import {
+  cls,
+  type StyleProps,
+  getThemeClassName,
+  useTheme,
+} from '@dopt/react-theme';
+
+export interface DismissIconProps
+  extends ComponentPropsWithRef<'button'>,
+    StyleProps {
+  onClick?: MouseEventHandler<HTMLButtonElement>;
+}
+
+const dismissIconClassName = `${classNameRoot}__dismiss-icon` as const;
+
+function TourItemDismissIcon(
+  props: DismissIconProps,
+  ref?: ForwardedRef<HTMLButtonElement>
+) {
+  const { theme: injectedTheme, className, onClick, ...restProps } = props;
+
+  const theme = useTheme(injectedTheme);
+
+  const icon = (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      strokeWidth="2"
+      stroke="currentColor"
+      fill="none"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    >
+      <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
+      <path d="M18 6l-12 12"></path>
+      <path d="M6 6l12 12"></path>
+    </svg>
+  );
+
+  return (
+    <button
+      onClick={onClick}
+      className={cls([
+        getThemeClassName({
+          theme,
+          className: [classes.dismissIcon(), theme],
+        }),
+        dismissIconClassName,
+        className,
+      ])}
+      {...restProps}
+      ref={ref}
+    >
+      {icon}
+    </button>
+  );
+}
+
+const DismissIcon = forwardRef(TourItemDismissIcon);
+export { DismissIcon, DismissIcon as TourItemDismissIcon };