+++ components/@dopt/react/tour/src/tour/footer.tsx
@@ -1,0 +1,46 @@
+import * as classes from '../styles';
+import { classNameRoot } from '../const';
+
+import {
+  type ForwardedRef,
+  forwardRef,
+  type ComponentPropsWithRef,
+} from 'react';
+
+import {
+  cls,
+  type StyleProps,
+  getThemeClassName,
+  useTheme,
+} from '@dopt/react-theme';
+
+export interface FooterProps
+  extends Omit<ComponentPropsWithRef<'footer'>, 'title'>,
+    StyleProps {}
+
+const footerClassName = `${classNameRoot}__footer` as const;
+
+function TourItemFooter(props: FooterProps, ref?: ForwardedRef<HTMLElement>) {
+  const { theme: injectedTheme, className, ...restProps } = props;
+
+  const theme = useTheme(injectedTheme);
+
+  return (
+    <footer
+      className={cls([
+        getThemeClassName({
+          theme,
+          className: [classes.footer(), theme],
+        }),
+        footerClassName,
+        className,
+      ])}
+      {...restProps}
+      ref={ref}
+    />
+  );
+}
+
+const Footer = forwardRef(TourItemFooter);
+
+export { Footer, Footer as TourItemFooter };